Improve index.html
I moved the selection variables outside of the playARound function, as suggested by the lesson. By doing so, it is now possible to call on those values as the arguments in a function call of playARound. You can see that is now how the result is displayed, "console.log(playARound(playerSelection, computerSelection))."

Much cleaner. Also improved the return strings for each scenario and included the empty string condition for the return "Game canceled. Please reload to try again."