Build game function and rework single round code
It became clear that, in order to have the prompt for user's choice show up when I wished it to, I needed to insert the prompt into the beginning of the code for the function playRound (which is the new name for what used to be playARound).
A single round will play just fine, which can be tested by uncommenting line 68 and commenting out all that comes after it.

Next step will be to see if the game function works to run 5 rounds and output the score after each, computing and declaring a winner at the end.